Finished

+++ app/Http/Controllers/HobbyController.php
@@ -5,17 +5,18 @@ namespace App\Http\Controllers;
 use App\Hobby;
 use App\Tag;
 use Illuminate\Http\Request;
-use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Session;
-
+use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyController extends Controller
-{   
+{
+
     public function __construct()
     {
         $this->middleware('auth')->except(['index', 'show']);
     }
-    
+
     /**
      * Display a listing of the resource.
      *
@@ -25,11 +26,14 @@ class HobbyController extends Controller
     {
         //$hobbies = Hobby::all();
         //$hobbies = Hobby::paginate(10);
+
         $hobbies = Hobby::orderBy('created_at', 'DESC')->paginate(10);
+
 
         return view('hobby.index')->with([
             'hobbies' => $hobbies
         ]);
+
     }
 
     /**
@@ -50,24 +54,24 @@ class HobbyController extends Controller
      */
     public function store(Request $request)
     {
+
         $request->validate([
             'name' => 'required|min:3',
-            'description' => 'required|min:5'
+            'description' => 'required|min:5',
+            'image' => 'mimes:jpeg,jpg,bmp,png,gif'
         ]);
-        
-        $hobby = New Hobby([
+
+        $hobby = new Hobby([
             'name' => $request['name'],
             'description' => $request['description'],
             'user_id' => auth()->id()
         ]);
         $hobby->save();
-        /*
-        return $this->index()->with(
-            [
-                'message_success' => "The hobby <b>" . $hobby->name . "</b> was created."
-            ]
-        );
-        */
+
+        if ($request->image) {
+            $this->saveImages($request->image, $hobby->id);
+        }
+
         return redirect('/hobby/' . $hobby->id)->with(
             [
                 'message_warning' => "Please assign some tags now."
@@ -87,13 +91,12 @@ class HobbyController extends Controller
         $usedTags = $hobby->tags;
         $availableTags = $allTags->diff($usedTags);
 
-
         return view('hobby.show')->with([
             'hobby' => $hobby,
             'availableTags' => $availableTags,
             'message_success' => Session::get('message_success'),
             'message_warning' => Session::get('message_warning')
-        ]); 
+        ]);
     }
 
     /**
@@ -104,8 +107,13 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         return view('hobby.edit')->with([
-            'hobby' => $hobby
+            'hobby' => $hobby,
+            'message_success' => Session::get('message_success'),
+            'message_warning' => Session::get('message_warning')
         ]);
     }
 
@@ -118,21 +126,30 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         $request->validate([
             'name' => 'required|min:3',
             'description' => 'required|min:5',
+            'image' => 'mimes:jpeg,jpg,bmp,png,gif'
         ]);
+
+        if ($request->image) {
+            $this->saveImages($request->image, $hobby->id);
+        }
 
         $hobby->update([
             'name' => $request['name'],
             'description' => $request['description'],
         ]);
 
-        return $this->index()->with(
-            [
-                'message_success' => "The hobby <b>" . $hobby->name . "</b> was updated."
-            ]
-        );
+        return $this->index()->withSuccess("The hobby <b>" . $hobby->name . "</b> was updated.");
+        //with(
+        //    [
+        //        'message_success' => "The hobby <b>" . $hobby->name . "</b> was updated."
+        //    ]
+        //);
     }
 
     /**
@@ -143,12 +160,50 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+
+        abort_unless(Gate::allows('delete', $hobby), 403);
+
         $oldName = $hobby->name;
         $hobby->delete();
+        return back()
+        ->withSuccess("New product was deleted");
+      
+    }
 
-        return $this->index()->with(
+    public function saveImages($imageInput, $hobby_id){
+
+        $image = Image::make($imageInput);
+        if ( $image->width() > $image->height() ) { // Landscape
+            $image->widen(1200)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg")
+                ->widen(400)->pixelate(12)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+            $image = Image::make($imageInput);
+            $image->widen(60)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+        } else { // Portrait
+            $image->heighten(900)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg")
+                ->heighten(400)->pixelate(12)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+            $image = Image::make($imageInput);
+            $image->heighten(60)
+                ->save(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+        }
+
+    }
+
+    public function deleteImages($hobby_id){
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_large.jpg");
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_thumb.jpg");
+        if(file_exists(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg"))
+            unlink(public_path() . "/img/hobbies/" . $hobby_id . "_pixelated.jpg");
+
+        return back()->with(
             [
-                'message_success' => "The hobby <b>" . $hobby->name . "</b> was deleted."
+                'message_success' => "The Image was deleted."
             ]
         );
     }
