+++ app/Http/Controllers/HobbyTagController.php
@@ -1,46 +1,58 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Tag;
-use App\Hobby;
-
 
 use Illuminate\Http\Request;
+use App\Tag;
+use App\Hobby;
+use Illuminate\Support\Facades\Gate;
 
-class HobbyTagController extends Controller
+
+class hobbyTagController extends Controller
 {
-    public function getFilteredHobbies($tag_id){
+    public function getFilteredHobbies($tag_id) {
+
         $tag = new Tag();
         $hobbies = $tag::findOrFail($tag_id)->filteredHobbies()->paginate(10);
 
         $filter = $tag::find($tag_id);
-       
+
 
         return view('hobby.index', [
             'hobbies' => $hobbies,
             'filter' => $filter
         ]);
+    }
+
+
+    public function attachTag($hobby_id, $tag_id) {
+
+        $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('conest_hobbyTag', $hobby)) {
+            abort(403, "That hobby do not belong to You!");
         }
 
-        public function attachTag($hobby_id, $tag_id){
-           
-            $hobby = Hobby::find($hobby_id);
-            $tag = Tag::find($tag_id);
-            $hobby->tags()->attach($tag_id);
-            return back()->with([
-                'message_success' => "The tag <b>" . $tag->name . "</b> was added."           
-                 ]);
-        }
+        $tag = Tag::find($tag_id);
+        $hobby->tags()->attach($tag_id);
+        return back()->with([
+            'message_success' => "The Tag <b>" . $tag->name . "</b> was added."
+        ]);
 
-        public function detachTag($hobby_id, $tag_id){
-            
-            $hobby = Hobby::find($hobby_id);
-            $tag = Tag::find($tag_id);
-            $hobby->tags()->detach($tag_id);
-            return back()->with([
-                 'message_success' => "The tag <b>" . $tag->name . "</b> was removed."           
-                ]);
-        }
+    }
 
+    public function detachTag($hobby_id, $tag_id) {
+
+        $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('conest_hobbyTag', $hobby)) {
+            abort(403, "That hobby do not belong to You!");
+        }
+        $tag = Tag::find($tag_id);
+        $hobby->tags()->detach($tag_id);
+        return back()->with([
+            'message_success' => "The Tag <b>" . $tag->name . "</b> was removed."
+        ]);
+    }
 
 }
