+++ app/Http/Controllers/TagController.php
@@ -4,11 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Session;
-
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index']);
+       
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -17,9 +23,10 @@ class TagController extends Controller
     public function index()
     {
         $tags = Tag::all();
+
         return view('tag.index')->with([
-        'tags' => $tags
-]);
+            'tags' => $tags
+        ]);
     }
 
     /**
@@ -42,15 +49,14 @@ class TagController extends Controller
     {
         $request->validate([
             'name' => 'required',
-            'style' => 'required'
+            'style' => 'required',
         ]);
-                
-        $tag = New Tag([
+
+        $tag = new Tag([
             'name' => $request['name'],
-            'style' => $request['style']
+            'style' => $request['style'],
         ]);
         $tag->save();
-        
         return $this->index()->with(
             [
                 'message_success' => "The tag <b>" . $tag->name . "</b> was created."
@@ -67,6 +73,8 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
+        
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -81,10 +89,14 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
+        
+        
         $request->validate([
             'name' => 'required',
             'style' => 'required',
         ]);
+
         $tag->update([
             'name' => $request['name'],
             'style' => $request['style'],
@@ -92,7 +104,7 @@ class TagController extends Controller
 
         return $this->index()->with(
             [
-                'message_success' => "The hobby <b>" . $tag->name . "</b> was updated."
+                'message_success' => "The tag <b>" . $tag->name . "</b> was updated."
             ]
         );
     }
@@ -105,12 +117,14 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        abort_unless(Gate::allows('delete', $tag), 403);
+        
+
         $oldName = $tag->name;
         $tag->delete();
-
         return $this->index()->with(
             [
-                'message_success' => "The tag <b>" . $tag->name . "</b> was deleted."
+                'message_success' => "The tag <b>" . $oldName . "</b> was deleted."
             ]
         );
     }
